pyproject.toml
[project]
name = "toy-diffusion"
version = "0.1.0"
description = "Educational DDPM for generating 32x32 binary pixel art of geometric shapes"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.24.0",
    "pillow>=9.0.0",
    "tqdm>=4.65.0",
    "tensorboard>=2.12.0",
    "matplotlib>=3.7.0",
]

[project.optional-dependencies]
modal = ["modal>=0.50.0"]
dev = ["pytest>=7.0.0", "black>=23.0.0", "ruff>=0.0.260"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]

[tool.black]
line-length = 100
